Simplify JSON schema with serde snake_case
