+++ caldir-core/src/event/status.rs
@@ -1,6 +1,7 @@
 use std::fmt;
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Default, serde::Serialize)]
+#[serde(rename_all = "snake_case")]
 pub enum Status {
     // CONFIRMED is the RFC 5545 default — an event with no STATUS line is
     // semantically confirmed, so absence and "confirmed" must round-trip the
