+++ caldir-core/src/event/visibility.rs
@@ -1,6 +1,7 @@
 use std::fmt;
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, serde::Serialize)]
+#[serde(rename_all = "snake_case")]
 pub enum Visibility {
     // RFC 5545 CLASS values. Absent CLASS maps to `None`, so no Default here.
     Public,
